Change tests for different tests per database
